Move ViewModelFactory to commons
